moved `MISSING` class to other file
