+++ interactions/api/models/misc.pyi
@@ -59,3 +59,5 @@ class Format:
     TIMESTAMP_LONG_DT: str
     TIMESTAMP_RELATIVE: str
     def stylize(self, format: str, **kwargs) -> str: ...
+
+class MISSING: ...
