+++ interactions/decor.py
@@ -1,10 +1,10 @@
 from typing import Any, Dict, List, Optional, Union
 
 from .api.models.guild import Guild
+from .api.models.misc import MISSING
 from .enums import ApplicationCommandType
 from .models.command import ApplicationCommand, Option
 from .models.component import Button, Component, SelectMenu
-from .models.misc import MISSING
 
 
 def command(
