+++ interactions/models/misc.py
@@ -184,9 +184,3 @@ class Interaction(DictSerializerMixin):
         self.channel_id = Snowflake(self.channel_id) if self._json.get("channel_id") else None
         self.member = Member(**self.member) if self._json.get("member") else None
         self.user = User(**self.user) if self._json.get("user") else None
-
-
-class MISSING:
-    """A pseudosentinel based from an empty object. This does violate PEP, but, I don't care."""
-
-    ...
